+++ CHANGELOG.md
@@ -26,11 +26,13 @@
 
 ## Misc
 
-- Use Shellcheck on tests/ files, changes for Shellcheck in tests/ (#368)
+- Use Shellcheck on tests/ files (#368)
+- Run Shellcheck on osx travis tests (#403)
 - Document SECRETS_VERBOSE and improve env var docs (#396)
 - Update CHANGELOG.md to mention fix for #281 in v0.2.5 (#311)
 - Add text explaining git-secret Style Guide and Development Philosophy
 - Upgrade bats-core to v1.1.0, import bats-core into vendor/bats-core (#377)
+- Use pgrep and kill to stop test gpg-agent subprocess (#376)
 
 ## Version 0.2.5